feat: now thats an efficient I/O stream :) !!!

+++ System/Peripherals/STM32F303R/uart2.hpp
@@ -12,8 +12,8 @@ public:
 protected:
     bool initUnsafe() override;
     bool deinitUnsafe() override;
-    bool txUnsafe(const std::uint8_t *pData, std::uint16_t size) override;
-    bool rxUnsafe(std::uint8_t *pData, std::uint16_t size) override;
+    bool txUnsafe(const std::uint8_t *pData, std::size_t size) override;
+    bool rxUnsafe(std::uint8_t *pData, std::size_t size) override;
 
     void operator=(Uart2 const &) = delete;
     Uart2(Uart2 const &) = delete;
