+++ System/Peripherals/STM32F303R/uart3.hpp
@@ -12,8 +12,8 @@ public:
 protected:
     bool initUnsafe() override;
     bool deinitUnsafe() override;
-    bool txUnsafe(const std::uint8_t *pData, std::uint16_t size) override;
-    bool rxUnsafe(std::uint8_t *pData, std::uint16_t size) override;
+    bool txUnsafe(const std::uint8_t *pData, std::size_t size) override;
+    bool rxUnsafe(std::uint8_t *pData, std::size_t size) override;
 
     void operator=(Uart3 const &) = delete;
     Uart3(Uart3 const &) = delete;
